requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# create_voters.py: 16
tqdm == 3.8.0

# locustfile.py: 4
locustio == 0.7.3

# create_voters.py: 5,6,7
pycrypto == 2.6.1

# setup_ledger.py: 7
fake_factory == 0.5.3

# locustfile.py: 5
# setup_ledger.py: 4,5
requests == 2.9.1

# setup_ledger.py: 29
click == 6.2
